test rework finished, (some fix in the test unit too)
