--- DApp/tests/runDemo.ts
@@ -1,5 +1,0 @@
-import UseCaseTest from "./UseCaseTest";
-
-// UseCaseTest.test_02();
-UseCaseTest.test_02B();
-
